feat(userSlice): user Slice relocated.
