+++ src/auth/pages/Login.tsx
@@ -5,7 +5,7 @@ import { useForm, type SubmitHandler } from 'react-hook-form'
 import { Copyright } from '../../components';
 import { loginUser, saveToken } from '../../api';
 import { useDispatch } from 'react-redux'
-import { login, setUser } from '../../redux/slice/userSlice';
+import { login, setUser } from '../../redux/slice/user/userSlice';
 
 interface Inputs {
   email: string
